requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
joblib>=1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
drf-spectacular==0.26.4
requests>=2.31.0
PyJWT>=2.8.0
